requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
SQLAlchemy==2.0.25
PyMySQL==1.1.0